Project created
This is the 5th exercise from a OOB C++/Java course that I took at uni. The history of commits here will be practically none-existant as I still haven't learned how to use git as it should be used.
Feel free to comment and clone my to your needs. It's all for free!

Pain(t) is based on Java language, although is uses JavaFX external libraries as it's base for GUI. To run the project from command line you need to first learn how to compile Java packages
and then combine it with compilation with external libraries.

As of the date of this commit no further development is planned.